+++ src/commands/deploy/promptToSaveDeployDefaults.ts
@@ -6,10 +6,10 @@
 import * as path from 'path';
 import { MessageItem } from "vscode";
 import { DialogResponses, IActionContext } from "vscode-azureextensionui";
-import * as constants from '../constants';
-import { SiteTreeItem } from '../explorer/SiteTreeItem';
-import { ext } from "../extensionVariables";
-import { getWorkspaceSetting, updateWorkspaceSetting } from "../vsCodeConfig/settings";
+import * as constants from '../../constants';
+import { SiteTreeItem } from '../../explorer/SiteTreeItem';
+import { ext } from "../../extensionVariables";
+import { getWorkspaceSetting, updateWorkspaceSetting } from "../../vsCodeConfig/settings";
 
 export async function promptToSaveDeployDefaults(context: IActionContext, node: SiteTreeItem, workspacePath: string, deployPath: string): Promise<void> {
     const defaultWebAppToDeploySetting: string | undefined = getWorkspaceSetting(constants.configurationSettings.defaultWebAppToDeploy, workspacePath);
